MemoryManagement&BugFix

+++ magic/verb.cpp
@@ -2,7 +2,7 @@
 
 using namespace Magic;
 
-Verb::Verb(Action action_, std::string n, Modifier e, Modifier c, Modifier d, double enemyCostMultiplier_, double allyEffectMultiplier_, double isSameMultiplier_)
+Verb::Verb(Action action_, std::string n, Modifier e, Modifier c, Modifier d, float enemyCostMultiplier_, float allyEffectMultiplier_, float isSameMultiplier_)
     :Word(n, e, c, d)
 {
     _action = action_;
@@ -17,17 +17,17 @@ WordType Verb::type(void) const
     return WordType::AVerb;
 }
 
-double Verb::enemyCostMultiplier(void) const
+float Verb::enemyCostMultiplier(void) const
 {
     return _enemyCostMultiplier;
 }
 
-double Verb::allyEffectMultiplier(void) const
+float Verb::allyEffectMultiplier(void) const
 {
     return _allyEffectMultiplier;
 }
 
-double Verb::isSameMultiplier(void) const
+float Verb::isSameMultiplier(void) const
 {
     return _isSameMultiplier;
 }
